fix: resolve 4 known bugs and add edge-case unit tests
Bugs fixed:
- EventPool: guard against double-dealloc (scan free list before accepting)
- Hysteresis: return level 0 for values below minimum threshold
- ADCReader: return 0.0f when min==max to prevent division by zero
- TCPServer: add nullptr check in read() and write()

Added 15 unit test files covering encoding round-trips, data structure
edge cases, serialization safety, algorithm boundaries, and bug
documentation. All tests compile with g++ -std=c++17 on host.

+++ src/ServiceLocator.h
@@ -1,6 +1,7 @@
 #ifndef PERSE_COMMON_SERVICELOCATOR_H
 #define PERSE_COMMON_SERVICELOCATOR_H
 
+#include <cstddef>
 #include <cstdint>
 #include <array>
 
@@ -11,16 +12,16 @@ class ServiceLocator {
 public:
 	template<typename T>
 	void set(Service service, T* ptr) {
-		services[static_cast<uint8_t>(service)] = static_cast<void*>(ptr);
+		services[static_cast<std::size_t>(service)] = static_cast<void*>(ptr);
 	}
 
 	template<typename T>
 	T* get(Service service) {
-		return static_cast<T*>(services[static_cast<uint8_t>(service)]);
+		return static_cast<T*>(services[static_cast<std::size_t>(service)]);
 	}
 
 private:
-	std::array<void*, static_cast<size_t>(Service::COUNT)> services{};
+	std::array<void*, static_cast<std::size_t>(Service::COUNT)> services{};
 };
 
 extern ServiceLocator Services;
